refactor(dao): BoardSelectCommand 패키지 이동
- View에서 사용하는 BoardSelectCommand가 DB를 조율하는 JanggiService까지 사용되는 경계 모호성 문제 해결
