+++ src/main/java/view/dto/BoardSelectCommand.java
@@ -1,4 +1,4 @@
-package domain.board;
+package view.dto;
 
 public record BoardSelectCommand(
         long select
